formatting

+++ overrides/olefile/0.47/default.nix
@@ -1,11 +1,10 @@
-{final, helpers, pkgs, ...}
-        : old: if (helpers.isWheel old)
-then {
-  postInstall = ''
-    rm $out/${final.python.sitePackages}/olefile/olefile2.py
-  '';
-}
-else {
-}
-
-        
+{ final, helpers, ... }:
+old:
+if (helpers.isWheel old) then
+  {
+    postInstall = ''
+      rm $out/${final.python.sitePackages}/olefile/olefile2.py
+    '';
+  }
+else
+  { }
